feat: restructure null index
