+++ lib/segment/src/index/field_index/null_index/immutable_null_index/read_ops.rs
@@ -1,12 +1,11 @@
-use std::path::PathBuf;
-
 use common::counter::hardware_accumulator::HwMeasurementAcc;
 use common::counter::hardware_counter::HardwareCounterCell;
 use common::types::PointOffsetType;
-use common::universal_io::UniversalRead;
+use common::universal_io::MmapFile;
 
-use super::read_ops::{self, NullIndexRead};
-use crate::common::flags::read_only_roaring_flags::ReadOnlyRoaringFlags;
+use super::super::read_ops::{self, NullIndexRead};
+use super::ImmutableNullIndex;
+use crate::common::flags::roaring_flags::RoaringFlags;
 use crate::common::operation_error::OperationResult;
 use crate::index::field_index::{
     CardinalityEstimation, PayloadBlockCondition, PayloadFieldIndexRead,
@@ -14,55 +13,33 @@ use crate::index::field_index::{
 use crate::index::query_optimization::optimized_filter::ConditionCheckerFn;
 use crate::types::{FieldCondition, PayloadKeyType};
 
-/// Read-only counterpart of [`MutableNullIndex`][1] / [`ImmutableNullIndex`][2].
-///
-/// All flags are loaded into in-memory roaring bitmaps on open. The backing
-/// storage is bound to [`UniversalRead`] only — no buffer, no flusher,
-/// no write path. Query logic (filter / cardinality / condition checker) is
-/// shared with the writable variant via [`read_logic`][3].
-///
-/// [1]: super::mutable_null_index::MutableNullIndex
-/// [2]: super::immutable_null_index::ImmutableNullIndex
-/// [3]: super::shared
-pub struct ReadOnlyNullIndex<S: UniversalRead> {
-    #[allow(dead_code)]
-    _base_dir: PathBuf,
-    storage: ReadOnlyStorage<S>,
-    total_point_count: usize,
-}
-
-struct ReadOnlyStorage<S: UniversalRead> {
-    /// Points which have at least one value
-    has_values_flags: ReadOnlyRoaringFlags<S>,
-    /// Points which have null values
-    is_null_flags: ReadOnlyRoaringFlags<S>,
-}
-
-impl<S: UniversalRead> NullIndexRead for ReadOnlyNullIndex<S> {
-    type Flags = ReadOnlyRoaringFlags<S>;
+impl NullIndexRead for ImmutableNullIndex {
+    type Flags = RoaringFlags<MmapFile>;
 
     fn has_values_flags(&self) -> &Self::Flags {
-        &self.storage.has_values_flags
+        self.0.has_values_flags()
     }
 
     fn is_null_flags(&self) -> &Self::Flags {
-        &self.storage.is_null_flags
+        self.0.is_null_flags()
     }
 
     fn total_point_count(&self) -> usize {
-        self.total_point_count
+        self.0.total_point_count()
     }
 
     fn telemetry_index_type(&self) -> &'static str {
-        "read_only_null_index"
+        "immutable_null_index"
     }
 }
 
-impl<S: UniversalRead> PayloadFieldIndexRead for ReadOnlyNullIndex<S> {
+impl PayloadFieldIndexRead for ImmutableNullIndex {
+    #[inline]
     fn count_indexed_points(&self) -> usize {
         self.indexed_points_count()
     }
 
+    #[inline]
     fn filter<'a>(
         &'a self,
         condition: &'a FieldCondition,
@@ -71,6 +48,7 @@ impl<S: UniversalRead> PayloadFieldIndexRead for ReadOnlyNullIndex<S> {
         Ok(read_ops::filter(self, condition))
     }
 
+    #[inline]
     fn estimate_cardinality(
         &self,
         condition: &FieldCondition,
@@ -79,6 +57,7 @@ impl<S: UniversalRead> PayloadFieldIndexRead for ReadOnlyNullIndex<S> {
         Ok(read_ops::estimate_cardinality(self, condition))
     }
 
+    #[inline]
     fn for_each_payload_block(
         &self,
         _threshold: usize,
